+++ tests/materials/material.helper.js
@@ -1,0 +1,45 @@
+const { gql } = require('@apollo/client');
+
+const createMaterial = async (material, operations) => {
+  const createdMaterial = await operations.mutate({
+    mutation: gql`
+      mutation($material: MaterialInput!) {
+        addMaterial(material: $material) {
+          ... on Material {
+            _id
+          }
+          ... on Error {
+            message
+          }
+        }
+      }
+    `,
+    variables: { material },
+  });
+
+  return createdMaterial.data.addMaterial._id;
+};
+const deleteMaterial = async (id, operations) => {
+  await operations.mutate({
+    mutation: gql`
+      mutation($id: ID!) {
+        deleteMaterial(id: $id) {
+          ... on Material {
+            _id
+          }
+          ... on Error {
+            message
+          }
+        }
+      }
+    `,
+    variables: {
+      id,
+    },
+  });
+};
+
+module.exports = {
+  deleteMaterial,
+  createMaterial,
+};
